requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
weaviate-client==4.5.4
google-generativeai
langchain
langchain-google-genai
langchain-community
langchain-tavily
duckduckgo-search
python-dotenv
google-cloud-secret-manager
protobuf>=4.21.0,<6.0.0 